% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clr.R
\name{clr}
\alias{clr}
\title{Centered Log-Ratio Transformation}
\usage{
clr(X)
}
\arguments{
\item{X}{numeric matrix or vector with all elements greater than or equal to 0.}
}
\description{
It calculates the centered log-ratio transformation of X
considering the rows as samples. If zero are present in X the function uses
the simplest replacement strategy replacing them with a constant value equal
to the 65\% of the detection limit of the sample (https://doi.org/10.1023/A:1023866030544).
}
