% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toy_model.R
\name{graph_toy_model}
\alias{graph_toy_model}
\title{Construct toy model network}
\usage{
graph_toy_model(obj, what)
}
\arguments{
\item{obj}{toy_model class object.}

\item{what}{normalization selected.}
}
\description{
construct an igraph object from toy_model object
}
