% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_toy_model.R
\name{plot.toy_model}
\alias{plot.toy_model}
\title{IN PROGRESS...}
\usage{
\method{plot}{toy_model}(x, what, layout = NULL, vertex.size = NULL, edge.width = NULL, ...)
}
\arguments{
\item{x}{obiect belong to toy_model S3 class}

\item{what}{...}

\item{layout}{coordinates of vertices of the network, layout_in_circle is set by default.}

\item{vertex.size}{...}

\item{edge.width}{...}

\item{...}{graphical paramaters to igraph plot}
}
\description{
IN PROGRESS...
}
