% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toy_model.R
\name{plot_toy_model}
\alias{plot_toy_model}
\alias{plot.toy_model}
\title{Custom plot for toy_model}
\usage{
\method{plot}{toy_model}(obj, what, ...)
}
\arguments{
\item{obj}{toy_model class object.}

\item{what}{normalization selected.}
}
\description{
plot ring network with custom graphical properties
}
