% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NorTA.R
\name{NorTA}
\alias{NorTA}
\title{Normal To Anything}
\usage{
NorTA(n, cor, dist, param = list())
}
\arguments{
\item{n}{number of observations.}

\item{cor}{correlation structure of resulting data; cor must be a number or a
correlation matrix. If it is a number it is understood that it is the
correlation between two variables.}

\item{dist}{characters indicate the chosen distribution.}

\item{param}{list containing the parameters of the distribution for each
dimension. All elements in the list must be named as the parameter name of
the distribution. If the elements have all lengths 1 the values are repeated
for all the dimensions.}
}
\description{
statistical technique to produce variables with correlation
structure and type of distribution chosen by the user.
}
