% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toy_model.R
\name{toy_model}
\alias{toy_model}
\title{Meta-genomic Toy Model}
\usage{
toy_model(
  n,
  cor,
  D,
  M,
  dist,
  param = list(),
  method = "pearson",
  seed = 42,
  force.positive = FALSE
)
}
\arguments{
\item{n}{number of observations.}

\item{cor}{correlation structure of resulting data; cor must be a number or a
correlation matrix. If it is a number it is understood that it is the
correlation between two variables.}

\item{D}{alternative choice to cor and indicates dimensionality. Produces a
correlation matrix of dimension D with by default two pairs of strongly
correlated variables, one positively and one negatively with correlations
values equal to +0.8 and -0.8. D cannot take values less than 5.}

\item{M}{magnification factor, real positive number that modify the
heterogeneity of the samples. In practice the function multiplies by a
variable by the value of M.}

\item{dist}{characters indicate the chosen distribution.}

\item{param}{list containing the parameters of the distribution for each
dimension. All elements in the list must be named as the parameter name of
the distribution. If the elements have all lengths 1 the values are repeated
for all the dimensions.}

\item{method}{type of correlation used. Possible choices are "pearson",
"kendall", "spearman" (default pearson).}

\item{seed}{random seed for reproducibility (default 42).}

\item{force.positive}{logical indicates when to force all elements of Y
to be positive number adding the minimum to all others (this passage does not
affect correlations).}
}
\value{
\code{toy_model} returns and object of class "toy_model" containing:
\itemize{
\item cor input correlation matrix.
\item normal generated gaussian data.
\item cor_normal correlation matrix of normal.
\item NorTA generated data of choosen distribution.
\item cor_NorTA correlation matrix of Y.
\item L1 relative abundances of NorTA.
\item cor_L1 correlation matrix of L1.
\item CLR matrix of abundances of clr transformed data from NorTA.
\item cor_CLR correlation matrix of CLR.
}
}
\description{
numerical simulations for meta-genomic data with user-chosen
correlation structure and sample heterogeneity.
}
